Save and restore app state when relevant callbacks are called.
This is neccessary to make sure state is not lost when Android kills
the app due to low memory when it is in the background.

It is also neccessary to avoid losing state (i. e. restarting the
game) when Android destroys and re-creates the activity when the
environment changes.  (For example, when I connect or disconnect my
Bluetooth keyboard.)